feat(gasOptimization): use uint256 when not for storage EVM requires the variable to be 256 so the compiler adds OPs for adding "0"s deployment improvement of 67930 gas (0.75%) execution cost improvement of 220 gas

+++ blockchain/contracts/DAO/governance/IGovernor.sol
@@ -128,7 +128,7 @@ abstract contract IGovernor is IERC6372 {
      * @notice module:core
      * @dev See {IERC6372}
      */
-    function clock() public view virtual returns (uint48);
+    function clock() public view virtual returns (uint256);
 
     /**
      * @notice module:core
